Move filter-related code into its own module
